+++ capstone/src/main/java/com/cognixia/project/repository/TodoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 import com.cognixia.project.model.Todo;
 
 @CrossOrigin("*")
-public interface TodoRepository extends JpaRepository<Todo, Integer> {
+public interface TodoRepository extends JpaRepository<Todo, Long> {
 }
